# docs/openapi.yaml
openapi: 3.0.0
info:
  title: CNS Express API
  version: 1.0.0
servers:
  - url: http://localhost:3000
tags:
  - name: sessions
  - name: streams
  - name: metrics
paths:
  /streams:
    get:
      tags:
        - streams
      summary: Get the list of available streams
      parameters:
        - in: cookie
          name: cns.connect.sid
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: Stream ID
                    name:
                      type: string
                      description: Stream name
                    description:
                      type: string
                      description: Stream description
                    src:
                      type: string
                      description: Stream source URL
                    resolutions:
                      type: array
                      items:
                        type: string
                      description: Supported resolutions

  /streams/{streamId}/stopped:
    post:
      tags:
        - streams
      summary: Handle stream stopped event
      parameters:
        - in: cookie
          name: cns.connect.sid
          schema:
            type: string
        - in: path
          name: streamId
          schema:
            type: string
          required: true
          description: Numeric ID of the user to get
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string

  /streams/{streamId}/started:
    post:
      tags:
        - streams
      summary: Handle stream started event
      parameters:
        - in: cookie
          name: cns.connect.sid
          schema:
            type: string
        - in: path
          name: streamId
          schema:
            type: string
          required: true
          description: Numeric ID of the user to get
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string

  /streams/{streamId}/views:
    post:
      tags:
        - streams
      summary: Handle stream view event
      parameters:
        - in: cookie
          name: cns.connect.sid
          schema:
            type: string
        - name: streamId
          in: path
          description: Stream ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resolution:
                  type: string
                  description: Stream resolution
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  clientIp:
                    type: string
                    description: Client IP address
                  userAgent:
                    type: string
                    description: User agent
                  sessionId:
                    type: string
                    description: Session ID
                  streamId:
                    type: string
                    description: Stream ID
                  resolution:
                    type: string
                    description: Stream resolution
                  views:
                    type: integer
                    description: Number of views
                  timestamp:
                    type: string
                    format: date-time
                    description: Timestamp of the response

  /streams/{streamId}/metrics:
    post:
      tags:
        - streams
      summary: Post steam metrics
      parameters:
        - in: cookie
          name: cns.connect.sid
          schema:
            type: string
        - name: streamId
          in: path
          description: Stream ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                trigger:
                  type: string
                  description: Trigger
                timestamp:
                  type: string
                  description: Timestamp
                screenSize:
                  type: object
                  properties:
                    width:
                      type: integer
                      description: Screen width
                    height:
                      type: integer
                      description: Screen height
                mediaLevel:
                  type: object
                  properties:
                    resolution:
                      type: string
                      description: Stream resolution
                    bandwidth:
                      type: integer
                      description: Stream bandwidth
                    level:
                      type: integer
                      description: Stream level
                    media:
                      type: string
                      description: Stream media
                streamedTime:
                  type: integer
                  description: Streamed time
                downloadedBytes:
                  type: integer
                  description: Downloaded bytes
                bufferings:
                  type: array
                  items:
                    type: object
                    properties:
                      timestamp:
                        type: string
                        description: Timestamp
                      videoTimestamp:
                        type: integer
                        description: Video timestamp
                      duration:
                        type: integer
                        description: Duration
                downloadRate:
                  type: integer
                  description: Download rate
                bandwidth:
                  type: integer
                  description: Bandwidth

      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  clientIp:
                    type: string
                    description: Client IP address
                  userAgent:
                    type: string
                    description: User agent
                  sessionId:
                    type: string
                    description: Session ID
                  streamId:
                    type: string
                    description: Stream ID
                  resolution:
                    type: string
                    description: Stream resolution
                  views:
                    type: integer
                    description: Number of views
                  timestamp:
                    type: string
                    format: date-time
                    description: Timestamp of the response

  /sessions:
    get:
      tags:
        - sessions
      summary: Get the list of all sessions ever created
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    description: Session ID
                  clientIp:
                    type: string
                    description: Client IP address
                  userAgent:
                    type: string
                    description: User agent
                  timestamp:
                    type: string
                    format: date-time
                    description: Timestamp of the response
    post:
      tags:
        - sessions
      summary: Obtain a new session, or get the current one if already exists
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    description: Session ID
                  clientIp:
                    type: string
                    description: Client IP address
                  userAgent:
                    type: string
                    description: User agent
                  timestamp:
                    type: string
                    format: date-time
                    description: Timestamp of the response
  /videos/{streamId}/master.m3u8:
    get:
      tags:
        - streams
      summary: Request .m3u8 file associated to a stream.
      parameters:
        - in: cookie
          name: cns.connect.sid
          schema:
            type: string
        - name: streamId
          in: path
          description: Video Name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    description: Session ID
                  clientIp:
                    type: string
                    description: Client IP address
                  userAgent:
                    type: string
                    description: User agent
                  timestamp:
                    type: string
                    format: date-time
                    description: Timestamp of the response